Divided Provisioner functionality in ProvisionerBase (parent) and Provisioner (child) classes.
